refactor: 검색 조건 패키지 수정
